Fix test.

+++ pkg/dotc1z/c1file_test.go
@@ -192,7 +192,7 @@ func TestC1ZDecoder(t *testing.T) {
 	require.Equal(t, dbFile, b.Bytes())
 	b.Reset()
 
-	require.GreaterOrEqual(t, n, 0)
+	require.GreaterOrEqual(t, n, int64(0))
 
 	// Test max size exact
 	//nolint:gosec // No risk of overflow because n is always >= 0.
@@ -206,7 +206,7 @@ func TestC1ZDecoder(t *testing.T) {
 	b.Reset()
 
 	// Test max size - 1
-	require.GreaterOrEqual(t, n, 1)
+	require.GreaterOrEqual(t, n, int64(1))
 	//nolint:gosec // No risk of overflow because n is always > 0.
 	d, err = NewDecoder(c1zf, WithDecoderMaxDecodedSize(uint64(n-1)))
 	require.NoError(t, err)
